+++ .env
@@ -1,6 +1,13 @@
 # Mandatory info for starting the app
-REACT_APP_MORALIS_APPLICATION_ID = mqTH3s5QdOWoIvQ2JpY3Kjgphc31cKnfuMyutIGu
-REACT_APP_MORALIS_SERVER_URL = https://nxbjd5xgdpfg.usemoralis.com:2053/server
+REACT_APP_MORALIS_APPLICATION_ID = BI5hagTDpD81Tc0hwFGn5kvW1nQgnjQKlDBrg0gT
+REACT_APP_MORALIS_SERVER_URL = https://pqrdt3y1tim2.usemoralis.com:2053/server
+
+# Polygon scan API Key
+API_KEY = 'QQWZ6TKJZVXCBDVHUYG5PMZH2EDZ2PN455'
+# Moralis Mumbai Endpoint 
+POLYGON_MUMBAI = 'https://speedy-nodes-nyc.moralis.io/b1a31628318e3734d3041852/polygon/mumbai'
+# Wallet Private key
+PRIVATE_KEY = '9ef07433abd17eb5479438636d216f16d0c8b3044070f2f28897710622f85bbb'
 
 # Optional info for connecting your localChain and Moralis Database
 moralisApiKey = xxxxx
